imported data persistence so when button presses we get something created a custom tableview cell implemented a tableview controller changed times to double redid the set for fahrenh eight created an image object model added a test to delete recent images.

+++ WeatherApp/WeatherApp/Models/Forecast.swift
@@ -93,8 +93,8 @@ struct DailyDatum: Codable {
     let time: Double
     let summary: String
     let icon: String
-    let sunriseTime: Int
-    let sunsetTime: Int
+    let sunriseTime: Double
+    let sunsetTime: Double
     let moonPhase: Double
     let precipIntensity: Double
     let precipIntensityMax: Double
